Live installer:     - Now you may pass one argument to "make_live.sh" to avoid       recreating ISO from scratch. Faster. Not for the 1st time.     - Patched patch to solve "Peez2" output bug.

+++ other_apps/live_test_system/make_live.sh
@@ -3,22 +3,26 @@
 SOURCESDIR=./sources
 ISONAME=gx2005.iso
 NAME="Custom Gx2005 Live"
-
 EXTRACTED_CD=./extracted_cd
+
 #Creamos la imagen
 echo Comprimiendo el sistema fuente...
 mkdir mnt.new
-sudo dd if=/dev/zero of=extracted_fs.new bs=1M count=2000
-yes s | sudo mkfs.ext2 extracted_fs.new
+
+if [[ $# == 0 ]]; then
+    sudo dd if=/dev/zero of=extracted_fs.new bs=1M count=2000
+    yes s | sudo mkfs.ext2 extracted_fs.new
+fi
+
 sudo mount -o loop,noatime,nodiratime extracted_fs.new mnt.new
 sudo rsync -avx --progress $SOURCESDIR/. mnt.new/.
 sudo umount mnt.new 
 rmdir mnt.new
 sudo create_compressed_fs extracted_fs.new 65536 > $EXTRACTED_CD/casper/filesystem.cloop
-sudo rm extracted_fs.new
 
-#Re-generamos el fichero .manifest
-#sudo chroot $SOURCESDIR dpkg-query -W --showformat='${Package} ${Version}\n' > $EXTRACTED_CD/casper/filesystem.manifest
+# if [[ $# > 0 ]]; then
+#     sudo rm extracted_fs.new
+# fi
 
 #Sumas md5
 echo Creando el md5sum.txt...
